Fix compiler warning in plexfluent.c
Reported by Intel(R) oneAPI DPC++/C++ Compiler 2023.1.0

petsc/src/dm/impls/plex/plexfluent.c:735:91: error: implicit conversion loses integer precision: 'PetscInt' (aka 'long') to 'int' [-Werror,-Wshorten-64-to-32]
          for (PetscInt c = 0; c < numCells; ++c) PetscCall(GetNumCellFaces(s.nd ? s.nd : ((PetscInt *)s.data)[c], &cellSizes[c], &cellTypes[c]));
                                                            ~~~~~~~~~~~~~~~               ^~~~~~~~~~~~~~~~~~~~~~~

cf. !8286

+++ src/dm/impls/plex/plexfluent.c
@@ -732,7 +732,7 @@ PetscErrorCode DMPlexCreateFluent(MPI_Comm comm, PetscViewer viewer, PetscBool i
           PetscCall(PetscInfo((PetscObject)viewer, "CASE: Found number of cells %" PetscInt_FMT "\n", numCells));
         } else {
           PetscCall(PetscMalloc2(numCells, &cellSizes, numCells, &cellTypes));
-          for (PetscInt c = 0; c < numCells; ++c) PetscCall(GetNumCellFaces(s.nd ? s.nd : ((PetscInt *)s.data)[c], &cellSizes[c], &cellTypes[c]));
+          for (PetscInt c = 0; c < numCells; ++c) PetscCall(GetNumCellFaces(s.nd ? s.nd : (int)((PetscInt *)s.data)[c], &cellSizes[c], &cellTypes[c]));
           PetscCall(PetscFree(s.data));
           PetscCall(PetscInfo((PetscObject)viewer, "CASE: Found number of cell faces %" PetscInt_FMT "\n", numCells && s.nd ? cellSizes[0] : 0));
         }
